Remove html error pages and use the ones from velum
This is needed because we need to revert the change in caasp-container-manifests
that was mounting the public directory from the host.

Mounting the public directory from the host, was "hidding" the assets
files present in the public directory and so our images were broken.

This was only happening in the iso and not in our dev environments
because in production we have the assets precompiled and we don't
compile them on the fly, since we don't have a js engine installed.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>